Split up ColBatch/ColVector into generic and non-generic parts
