+++ src/main/scala/com/nec/arrow/colvector/GenericColBatch.scala
@@ -1,0 +1,5 @@
+package com.nec.arrow.colvector
+
+final case class GenericColBatch[Data](numRows: Int, cols: List[Data]) {
+  def map[To](f: Data => To): GenericColBatch[To] = copy(cols = cols.map(f))
+}
