+++ src/main/scala/com/nec/ve/package.scala
@@ -1,0 +1,6 @@
+package com.nec
+
+package object ve {
+  type VeColBatch = colvector.VeColBatch
+  val VeColBatch = colvector.VeColBatch
+}
